feat: alert 창에 sweetalert2 적용

+++ src/pages/Profile.jsx
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from 'react'
 import './Profile.css';
 import axios from 'axios';
 import { useNavigate } from 'react-router-dom';
+import Swal from 'sweetalert2'
 
 const Profile = () => {
 
@@ -24,38 +25,53 @@ const Profile = () => {
     const handleTabChange = tab => setActiveTab(tab);
 
     // 닉네임 수정 아이콘 클릭 여부 변경 함수
-    const handleNicknameClick = click => setclickNickname(click);
+    const handleNicknameClick = () => setclickNickname(!clickNickname);
+
+    // Bearer Token 가져오기
+    const token = sessionStorage.getItem("accessToken");
+    console.log("profile token : ", token);
 
     // 닉네임 수정 버튼 클릭 시 실행 함수
     const updateNickname = async () => {
+        if(!nickname.trim()) {
+            Swal.fire({
+                icon: "warning",
+                title: "닉네임 입력 필수 !",
+                text: "닉네임을 입력하세요.",
+            });
+            return;
+        }
+            
         // nickname 값이 null이 아니면 수정 요청
-        if(nickname) {
-            try {
-                const serverResponse = await axios.patch("http://localhost:8080/members/me/edit", {
-                    nickname: nickname
-                });
+        try { 
+            const serverResponse = await axios.patch("http://localhost:8080/members/me/edit", { nickname });
     
-                // 닉네임 변경 요청 성공 시
-                if(serverResponse.status === 200) 
-                    console.log("닉네임 변경 성공");
+            // 닉네임 변경 요청 성공 시
+            if(serverResponse.status === 200) {
+                Swal.fire({
+                    icon: "success",
+                    title: "닉네임 변경 성공 !",
+                    text: "닉네임을 성공적으로 변경하였습니다."
+                });
+                // 닉네임 입력 창 닫기
+                setclickNickname(false);
+                // 닉네임 변경 후, 다시 사용자 정보 불러옴
+                getUserInfo();
             }
-            catch (error) {
-                console.error("닉네임 변경 실패 : ", error);
-            }  
         }
-        else {
-            alert("닉네임을 입력하세요 !");
-        }
+        catch (error) {
+            console.error("닉네임 변경 실패 : ", error);
+        }  
     }
+ 
 
     // 사용자 정보 요청 함수
     const getUserInfo = async () => {
-        const token = localStorage.getItem("accessToken");
         try {
             const serverResponse = await axios.get("http://localhost:8080/members/me", {
                 headers: {
                     Authorization: `Bearer ${token}`
-                }
+                }, withCredentials: true
             });
 
             if(serverResponse.data.success)
@@ -66,7 +82,7 @@ const Profile = () => {
         } 
     }
 
-    // 찜 목록 요청 함수
+    // 찜한 영화 목록 요청 함수
     const getLikeList = async () => {
         try {
             const serverResponse = await axios.get("http://localhost:8080/members/list");
@@ -84,23 +100,33 @@ const Profile = () => {
     
     // 회원 탈퇴 함수
     const deleteUser = async () => {
-        alert("계정을 삭제하시겠습니까 ?");
+        const result = await Swal.fire({
+            title: "정말 탈퇴하시겠습니까?",
+            text: "탈퇴 후 계정 복구가 불가능합니다.",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#d33",
+            cancelButtonColor: "#3085d6",
+            confirmButtonText: "네, 탈퇴할래요",
+            cancelButtonText: "취소"
+        });
 
-        const token = localStorage.getItem("accessToken");
-
-        try {
-            const serverResponse = await axios.delete("http://localhost:8080/members/withdrawl", {
-                headers: {
-                    Authorization: `Bearer ${token}`
-                }
-            });
-
-            if(serverResponse.data.success)
-                setUserInfo(serverResponse.data.data);
+        if(result.isConfirmed) {
+            try {
+                await axios.delete("http://localhost:8080/members/withdrawl", {
+                    headers: {
+                        Authorization: `Bearer ${token}`
+                    }, withCredentials: true 
+                });
+    
+                // 회원 탈퇴 시, 로컬 스토리지에서 토큰 삭제 후 로그인 페이지로 이동
+                sessionStorage.removeItem("accessToken");
+                navigate('/');
+            }
+            catch (error) {
+                console.error("회원 탈퇴 실패 : ", error);
+            } 
         }
-        catch (error) {
-            console.error("회원 탈퇴 실패 : ", error);
-        } 
     }
 
     // activeTab 변경될 때만 useEffect 호출
@@ -112,7 +138,6 @@ const Profile = () => {
         else if(activeTab === 'preference')
             getLikeList();
     }, [activeTab])
-
 
 
     return (
@@ -145,7 +170,7 @@ const Profile = () => {
                                     <div><strong>닉네임</strong></div>
                                     <div>홍길동</div>
                                     <img className='profile-editIcon' src='./src/assets/editIcon.png'
-                                         onClick={() => handleNicknameClick(!clickNickname)}/><br />
+                                         onClick={handleNicknameClick}/><br />
                                         {clickNickname ? (
                                             <>
                                                 <div><strong>닉네임 변경</strong></div>
